# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NER-LogiSense Intelligence Platform",
    "description": "\n# NER-LogiSense Backend API\n### AI-powered Smart Logistics and Accessibility Intelligence Platform for India's North Eastern Region (NER)\n\n**Core Capabilities**:\n- Real-time road, bridge, and transport accessibility monitoring\n- Landslide and flood disruption prediction using environmental feature models\n- AI-based multi-criteria route optimization ($Cost = \\alpha \\cdot T + \\beta \\cdot R + \\gamma \\cdot C$)\n- High-throughput GPS tracking of vehicles carrying medical oxygen, cryo-vaccines, foodgrains, and relief supplies\n- Offline-first field hazard reporting with idempotency and conflict resolution\n- Multilingual automated alerts across 9 regional languages\n- WebSocket telemetry streaming\n    ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/incidents": {
      "post": {
        "tags": [
          "Incidents & Field Reporting"
        ],
        "summary": "Create Incident",
        "description": "Report an incident (landslide, flood, bridge closure, congestion).",
        "operationId": "create_incident_api_v1_incidents_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Incidents & Field Reporting"
        ],
        "summary": "List Incidents",
        "description": "List all incidents with optional severity and status filters.",
        "operationId": "list_incidents_api_v1_incidents_get",
        "parameters": [
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          },
          {
            "name": "type_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type Filter"
            }
          },
          {
            "name": "severity_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity Filter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IncidentResponse"
                  },
                  "title": "Response List Incidents Api V1 Incidents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/{id}": {
      "get": {
        "tags": [
          "Incidents & Field Reporting"
        ],
        "summary": "Get Incident By Id",
        "description": "Get single incident details.",
        "operationId": "get_incident_by_id_api_v1_incidents__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Incidents & Field Reporting"
        ],
        "summary": "Update Incident",
        "description": "Update incident status or estimated clearance.",
        "operationId": "update_incident_api_v1_incidents__id__patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/{id}/verify": {
      "post": {
        "tags": [
          "Incidents & Field Reporting"
        ],
        "summary": "Verify Incident",
        "description": "Official verification by District Officer or BRO/PWD Chief.",
        "operationId": "verify_incident_api_v1_incidents__id__verify_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentVerifyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Verify Incident Api V1 Incidents  Id  Verify Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/{id}/resolve": {
      "post": {
        "tags": [
          "Incidents & Field Reporting"
        ],
        "summary": "Resolve Incident",
        "description": "Mark an incident resolved after debris clearance or water recession.",
        "operationId": "resolve_incident_api_v1_incidents__id__resolve_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Resolve Incident Api V1 Incidents  Id  Resolve Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/field-reports": {
      "post": {
        "tags": [
          "Incidents & Field Reporting"
        ],
        "summary": "Submit Field Report",
        "description": "Submit geo-tagged field report with offline UUID support.",
        "operationId": "submit_field_report_api_v1_field_reports_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FieldReportCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/field-reports/sync": {
      "post": {
        "tags": [
          "Incidents & Field Reporting"
        ],
        "summary": "Sync Offline Reports",
        "description": "Batch synchronize field reports queued during network blackout.",
        "operationId": "sync_offline_reports_api_v1_field_reports_sync_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/FieldReportCreate"
                },
                "type": "array",
                "title": "Payload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Sync Offline Reports Api V1 Field Reports Sync Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/media/upload": {
      "post": {
        "tags": [
          "Incidents & Field Reporting"
        ],
        "summary": "Upload Media File",
        "description": "Direct upload for photos/videos of road damages and landslides.",
        "operationId": "upload_media_file_api_v1_media_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_media_file_api_v1_media_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/media/{id}/presigned-url": {
      "get": {
        "tags": [
          "Incidents & Field Reporting"
        ],
        "summary": "Get Presigned Url",
        "description": "Obtain secure signed URL for S3/MinIO direct upload.",
        "operationId": "get_presigned_url_api_v1_media__id__presigned_url_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "filename",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "hazard.jpg",
              "title": "Filename"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Presigned Url Api V1 Media  Id  Presigned Url Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_media_file_api_v1_media_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_media_file_api_v1_media_upload_post"
      },
...
      "FieldReportCreate": {
        "properties": {
          "client_uuid": {
            "type": "string",
            "title": "Client Uuid"
          },
          "incident_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Incident Id"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lon": {
            "type": "number",
            "title": "Lon"
          },
          "accuracy_m": {
            "type": "number",
            "title": "Accuracy M",
            "default": 5.0
          },
          "disruption_type": {
            "type": "string",
            "title": "Disruption Type"
          },
          "severity": {
            "type": "string",
            "title": "Severity",
            "default": "CRITICAL"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "media_files": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Media Files"
          },
          "offline_created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Offline Created At"
          },
          "device_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id"
          }
        },
        "type": "object",
        "required": [
          "client_uuid",
          "lat",
          "lon",
          "disruption_type",
          "description",
          "offline_created_at"
        ],
        "title": "FieldReportCreate"
      },
      "FieldReportResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "client_uuid": {
            "type": "string",
            "title": "Client Uuid"
          },
          "incident_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Incident Id"
          },
          "reporter_id": {
            "type": "string",
            "title": "Reporter Id"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lon": {
            "type": "number",
            "title": "Lon"
          },
          "accuracy_m": {
            "type": "number",
            "title": "Accuracy M"
          },
          "disruption_type": {
            "type": "string",
            "title": "Disruption Type"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "media_files": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Media Files"
          },
          "offline_created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Offline Created At"
          },
          "synced_at": {
            "type": "string",
            "format": "date-time",
            "title": "Synced At"
          },
          "sync_status": {
            "type": "string",
            "title": "Sync Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "client_uuid",
          "reporter_id",
          "lat",
          "lon",
          "accuracy_m",
          "disruption_type",
          "severity",
          "description",
          "offline_created_at",
          "synced_at",
          "sync_status"
        ],
        "title": "FieldReportResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IncidentCreate": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "severity": {
            "type": "string",
            "title": "Severity",
            "default": "HIGH"
          },
          "location_desc": {
            "type": "string",
            "title": "Location Desc"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lon": {
            "type": "number",
            "title": "Lon"
          },
          "road_segment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Road Segment Id"
          },
          "district_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "District Id"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "estimated_clearance_hours": {
            "type": "number",
            "title": "Estimated Clearance Hours",
            "default": 4.0
          }
        },
        "type": "object",
        "required": [
          "type",
          "location_desc",
          "lat",
          "lon",
          "description"
        ],
        "title": "IncidentCreate"
      },
      "IncidentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "location_desc": {
            "type": "string",
            "title": "Location Desc"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lon": {
            "type": "number",
            "title": "Lon"
          },
          "road_segment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Road Segment Id"
          },
          "district_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "District Id"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "reported_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reported By"
          },
          "verified_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified By"
          },
          "estimated_clearance_hours": {
            "type": "number",
            "title": "Estimated Clearance Hours"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "severity",
          "location_desc",
          "lat",
          "lon",
          "description",
          "status",
          "estimated_clearance_hours",
          "created_at"
        ],
        "title": "IncidentResponse"
      },
      "IncidentUpdate": {
        "properties": {
          "severity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "estimated_clearance_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Clearance Hours"
          }
        },
        "type": "object",
        "title": "IncidentUpdate"
      },
      "IncidentVerifyRequest": {
        "properties": {
          "verified": {
            "type": "boolean",
            "title": "Verified"
          },
          "remarks": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remarks"
          },
          "updated_severity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated Severity"
          }
        },
        "type": "object",
        "required": [
          "verified"
        ],
        "title": "IncidentVerifyRequest"
      },
...
      "MediaUploadResponse": {
        "properties": {
          "media_id": {
            "type": "string",
            "title": "Media Id"
          },
          "file_url": {
            "type": "string",
            "title": "File Url"
          },
          "checksum_sha256": {
            "type": "string",
            "title": "Checksum Sha256"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "virus_scanned": {
            "type": "boolean",
            "title": "Virus Scanned"
          }
        },
        "type": "object",
        "required": [
          "media_id",
          "file_url",
          "checksum_sha256",
          "size_bytes",
          "virus_scanned"
        ],
        "title": "MediaUploadResponse"
      },
...
    }
  }
}